+++ code/sspp-assignment/cuda/ELLPACKCudaSolver.cu
@@ -8,9 +8,9 @@ namespace sspp {
   namespace cuda {
     namespace ellpack {
       template<typename VALUE_TYPE>
-      __global__ void EllpackKernel(unsigned rows,
-                                    unsigned max_row_non_zeros,
-                                    unsigned *JA,
+      __global__ void EllpackKernel(unsigned long long rows,
+                                    unsigned long long max_row_non_zeros,
+                                    unsigned long long *JA,
                                     VALUE_TYPE *AS,
                                     VALUE_TYPE *b,
                                     VALUE_TYPE *x) {
@@ -18,9 +18,9 @@ namespace sspp {
 
         if(row < rows) {
           VALUE_TYPE dot = VALUE_TYPE(0);
-          for(unsigned i = 0; i < max_row_non_zeros; i++) {
+          for(unsigned long long i = 0; i < max_row_non_zeros; i++) {
             auto index = row * max_row_non_zeros + i;
-            unsigned col = JA[index];
+            unsigned long long col = JA[index];
             VALUE_TYPE val = AS[index];
             if(val != 0)
               dot += val * b[col];
@@ -31,8 +31,8 @@ namespace sspp {
       }
 
       void SetUpWrapper(int& device_id,
-                        unsigned& thread_blocks,
-                        unsigned& threads_per_block) {
+                        unsigned long long& thread_blocks,
+                        unsigned long long& threads_per_block) {
         device_id = gpuGetMaxGflopsDeviceId();
         cudaSetDevice(device_id);
         cudaDeviceSetCacheConfig(cudaFuncCachePreferL1);
@@ -46,23 +46,23 @@ namespace sspp {
       template <typename VALUE_TYPE>
       void LoadELLPACK(common::ELLPACK<VALUE_TYPE>& ellpack,
                        std::vector<VALUE_TYPE>& vector,
-                       unsigned** d_column_indices,
+                       unsigned long long** d_column_indices,
                        VALUE_TYPE** d_values,
                        VALUE_TYPE** d_vector,
                        VALUE_TYPE** d_output) {
         checkCudaErrors(cudaMalloc(d_vector, sizeof(VALUE_TYPE) * vector.size()));
         checkCudaErrors(cudaMalloc(d_output, sizeof(VALUE_TYPE) * ellpack.GetRows()));
         checkCudaErrors(cudaMalloc(d_values, sizeof(VALUE_TYPE) * ellpack.GetValues().size()));
-        checkCudaErrors(cudaMalloc(d_column_indices, sizeof(unsigned) * ellpack.GetColumnIndices().size()));
+        checkCudaErrors(cudaMalloc(d_column_indices, sizeof(unsigned long long) * ellpack.GetColumnIndices().size()));
 
         checkCudaErrors(cudaMemcpy(*d_vector, &vector[0], sizeof(VALUE_TYPE) * vector.size(), cudaMemcpyHostToDevice));
         checkCudaErrors(cudaMemset(*d_output, 0, sizeof(VALUE_TYPE) * ellpack.GetRows()));
         checkCudaErrors(cudaMemcpy(*d_values, &ellpack.GetValues()[0], sizeof(VALUE_TYPE) * ellpack.GetValues().size(), cudaMemcpyHostToDevice));
-        checkCudaErrors(cudaMemcpy(*d_column_indices, &ellpack.GetColumnIndices()[0], sizeof(unsigned) * ellpack.GetColumnIndices().size(), cudaMemcpyHostToDevice));
+        checkCudaErrors(cudaMemcpy(*d_column_indices, &ellpack.GetColumnIndices()[0], sizeof(unsigned long long) * ellpack.GetColumnIndices().size(), cudaMemcpyHostToDevice));
       }
 
       template <typename VALUE_TYPE>
-      void ReleaseELLPACK(unsigned* d_column_indices,
+      void ReleaseELLPACK(unsigned long long* d_column_indices,
                           VALUE_TYPE* d_values,
                           VALUE_TYPE* d_vector,
                           VALUE_TYPE* d_output) {
@@ -80,11 +80,11 @@ namespace sspp {
       }
 
       template <typename VALUE_TYPE>
-      void SolveELLPACK(unsigned thread_blocks,
-                        unsigned threads_per_block,
-                        unsigned rows,
-                        unsigned max_row_non_zeros,
-                        unsigned* d_column_indices,
+      void SolveELLPACK(unsigned long long thread_blocks,
+                        unsigned long long threads_per_block,
+                        unsigned long long rows,
+                        unsigned long long max_row_non_zeros,
+                        unsigned long long* d_column_indices,
                         VALUE_TYPE* d_values,
                         VALUE_TYPE* d_vector,
                         VALUE_TYPE* d_output) {
@@ -93,7 +93,7 @@ namespace sspp {
       }
 
       template <typename VALUE_TYPE>
-      std::vector<VALUE_TYPE> GetResult(VALUE_TYPE* d_output, unsigned size) {
+      std::vector<VALUE_TYPE> GetResult(VALUE_TYPE* d_output, unsigned long long size) {
         std::vector<VALUE_TYPE> result(size);
         checkCudaErrors(cudaMemcpy(&result[0], d_output, sizeof(VALUE_TYPE) * size, cudaMemcpyDeviceToHost));
         return result;
